imx_usb: moved struct mach_id to usb specific part
Moved struct mach_id to usb specific part. In the USB case, we have
the chance to actually detect what type of machine is connected.
The imx_usb.conf file represents the database for this machines,
however in the UART case we don't have a reliable solution to
detect the machine. The plan is to ask the user to select the right
machine configuration file. Hence, move this to the USB part and
pass the filename only.