imx_usb/imx_uart: new argument parsing, add help text
Implement a more sophisticated argument parsing. The UART variant
checks for mandatory arguments and allows optional arguments as
well. Move the verify to the optional arguments at the beginning.
Add configuration directory parameter for the USB variant and a
option to disable CTS/RTS flow control for the UART variant.

The Job/Work parsing itself is still shared and available at the
end of the arguments.  Multiple Job parsing generally works,
however execution seems to be broken right now.

Also implement a help text outlining the available options and
mandatory arguments (if any).