Actually allow changing the configuration file location (sysconfdir).
Binaries would always try to access the configuration file in /etc,
regardless of the specified prefix or sysconfdir.

Add a preprocessor define and use it when accessing the configuration
file.  This way, a user may install it in its home directory and
have the binary work out of the box.