Add no_clear_boot_data option
I'm using this option to boot a HAB-signed image. I'm using two entries
in my work configuration:

  u-boot.imx:dcd
  u-boot-nodcd.imx:no_clear_boot_data,jump header

The first one is an u-boot image that is just used to read out a DCD.
The second one is an u-boot image that is signed with a DCD pointer set
to zero.

Setting the boot_data pointer also zero isn't possible because it's
required to read the headers. So it would be necessary to generate an
image that was signed with the boot_data pointer set to zero but
contains valid boot data.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>