imx_sdp: Change DCD OCRAM address to match Freescale tools
With the current OCRAM address for DCD used in imx_usb, a bootloader
would need to be signed twice with different configurations to
support both imx_usb and the Freescale MfgTool when using High
Assurance Boot (HAB).

Updated the OCRAM address where the DCD is loaded so if a bootloader
is signed, it can be loaded with both Freescale tools (MfgTool) and
the imx_usb program without needing to be signed twice.