deduplicate device initialization code
So far the function open_vid_pid did a lot of initialization which
is already done as part of the device initialization in main. By
moving the job handling part above the retry code, we can jump
out of the loop and reuse the initialization code. Once we get to
the loop again, we will continue with the next job.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>