simplify and document work processing
When using command line, the current code does reuses the same job
to execute the plugin and then download the actual image. If there
is only a image, it also makes sure that it does not get executed
twice.

To further refactor the code, we would like to handle a struct sdp_work
independent from its origin. This change does so by getting rid of
the reference to cmd_head. It comes with a slight behavior change:
If one specifies a single job with a plugin, it will also download
the image. Since this is most likely what you intended anyway, the
behavior change is considered acceptable.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>