move USB work handling in its own function
Move the actual work handling in a seperate function. This makes
main smaller and allows to reuse more libusb boilerplate code.

It also moves the retry handling out of the while loop and to the
parent function. With that we have a do_work function which does
a single attempt of trying to run the given work list.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>