Improve error handling around DoIRomDownload
Print a message in case DoIRomDownload fails. Also do not return
an error in DoIRomDownload in case status fails after a jump:
If the jump succeeds, the Boot ROM won't answer to the status
call, hence an error is expected here.

We still want to do a status call, since it might be that the
jump failed.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>