Improve error handling around DoIRomDownload
Print a message in case DoIRomDownload fails. Also do not return
an error in DoIRomDownload in case status fails after a jump:
If the jump succeeds, the Boot ROM won't answer to the status
call, hence an error is expected here. We still do do a status
call so we can determine the error code in case the jump failed.

If DoIRomDownload returns an error code, e.g. because a file is not
accessible, the following status command should not overwrite the
error so that we have the chance to exit.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>