Transfer always specified report 2 length
The Windows HIDAPI is very strict when it comes to package lengths:
The API always expect the length specified by the device and rejects
the transfer otherwise. The report size is specified by the HID
Report Descriptor. The i.MX 6 SoC reports a report 2 size of 1024.
Currently, imx_usb reads the length from the device specific config
files and stores it in max_transfer. This change uses max_transfer
for report 2 when using the HID API.

With that change, we might end up transfer up to 1023 bytes more
than actually necessary. Make sure that DoIRomDownload does not
return an error in this case. Also don't print an error message
in case the transfer transmits more than requested (which happens
typically at the end of every file).

The boot ROM handles too long transfers just fine. Also NXP's
sb_loader transfers always complete reports. The change has been
successfully tested on Linux and Windows.

Given that the device reports the size, it is probably also "the
right thing to do"...

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>