imx_sdp: extract response handling
This allows for better error handling and makes sure we always
use a empty buffer to read into.

Note that this changes endianness of the 4 byte response (security
configuration/status). Since the protocol otherwise is big endian
this is the right thing to do. It used to work so far since the
results which have been checked were all symetric (0x128a8a12UL and
0x88888888UL).

Also show the security configuration since we transmit it already.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>