remove default load address
Default load address is not used when specifying using arguments
since header is set to J_HEADER. So the only situation when
load address is used is when a config file is specified. But when
using config files the user should use the regular load command
to download a file, e.g.:
file:load nnn

This fixes an issue when trying to load only DCD/plugin using a
config file. In that case imx_usb_loader is currently downloading
the file despite there is no such line in the config file.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>