avoid calling do_download twice if plugin is used
This is a long standing oddity of the current way how a work
item gets processed: Header features such as DCD gets processed
right in process_header. However, if a plugin is found, the code
returns all the way back to do_work, only to call with the same
work item again.

This is weired, complex and unnecessary. Just make sure that the
plugin gets downloaded/jumped to right in process_header.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>