imx_uart: fix option parsing on ARM and other unsigned-char systems
imx_uart assigns the return value of getopt_long() to a char variable.

On systems that default to unsigned char (e.g. arm, aarch64, powerpc,
s390) this causes the termination return value -1 to be read as 0xff and
the option parsing loop never terminates, causing the program to get
immediately stuck.

Fix that by using an int which is the actual return type of
getopt_long().

imx_usb already correctly uses int.

Fixes #116.

Signed-off-by: Anssi Hannula <anssi.hannula@bitwise.fi>